+++ DungeonFarming/DungeonFarming/DataBaseServices/GameDb/MasterDataDTO/ItemDefine.cs
@@ -2,7 +2,7 @@
 {
     public class ItemDefine
     {
-        public Int16 item_code { get; set; }
+        public Int64 item_code { get; set; }
         public String item_name { get; set; }
         public Int16 attribute { get; set; }
         public Int64 sell { get; set; }
